# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  v1.ApplicationFormDTO:
    properties:
      formSchema:
        additionalProperties: true
        type: object
    type: object
...
  v1.DepartmentDTO:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  v1.LevelDTO:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  v1.LocationDTO:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  v1.PublicVacancyDetails:
    properties:
      department:
        $ref: '#/definitions/v1.DepartmentDTO'
      id:
        type: string
      important:
        type: boolean
      info:
        type: string
      level:
        $ref: '#/definitions/v1.LevelDTO'
      location:
        $ref: '#/definitions/v1.LocationDTO'
      logoUrl:
        type: string
      name:
        type: string
      photoUrl:
        type: string
      priority:
        type: integer
    type: object
  v1.PublicVacancyItem:
    properties:
      department:
        $ref: '#/definitions/v1.DepartmentDTO'
      id:
        type: string
      important:
        type: boolean
      level:
        $ref: '#/definitions/v1.LevelDTO'
      location:
        $ref: '#/definitions/v1.LocationDTO'
      logoUrl:
        type: string
      name:
        type: string
      priority:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a job website.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Job Website API
  version: "1.0"
paths:
...
  /vacancies:
    get:
      consumes:
      - application/json
      description: Get a list of active vacancies
      parameters:
      - description: filter by important vacancies
        in: query
        name: important
        type: boolean
      - description: sort by priority (asc/desc)
        in: query
        name: sort
        type: string
      - description: search by vacancy name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.PublicVacancyItem'
            type: array
      summary: Get a list of active vacancies
      tags:
      - vacancies
  /vacancies/{id}:
    get:
      consumes:
      - application/json
      description: Get a vacancy by ID
      parameters:
      - description: Vacancy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PublicVacancyDetails'
      summary: Get a vacancy by ID
      tags:
      - vacancies
  /vacancies/{id}/form:
    get:
      consumes:
      - application/json
      description: Get the application form for a vacancy
      parameters:
      - description: Vacancy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ApplicationFormDTO'
      summary: Get the application form for a vacancy
      tags:
      - vacancies
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"